+++ CHANGELOG.md
@@ -34,6 +34,10 @@
 
 
 ### Bug Fixes
+* **o-combo**: Fixing change event emission ([a648c8b](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/a648c8b)) Closes [#152](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/152)
+* **o-form**: Fixing close dialog on ESC key event ([f789a64](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/f789a64))
+
+
 ## 2.1.2 (2018-04-03)
 ### PEER-DEPENDENCY UPDATES ###
 * **Updated**:  core-js@2.5.3